// src/app/shared/model/catalogo.ts
export interface Catalogo {
    cpIdCatalogo?: number;
    cpCodigoCatalogo?: string;
    cpCampo1?: string;
    cpCampo2?: string;
    cpCampo3?: string;
    cpCampo4?: string;
    cpCampo5?: string;
    cpCampo6?: string;
    cpCampo7?: string;
    cpCampo8?: string;
    cpCampo9?: string;
    cpCampo10?: string;
    cpCampo11?: number;
    cpestado?: boolean;
}